Added mpris tunecontroller support based on Matthieu Volat's code
I removed the check for the service name, so that the corresponging signals from every mpris music player are recognized.
The downside is, that even signals from players, that don't use the (org.mpris.<name>) service name are accepted. But for now I think it's still better than to require a seperate plugin for every player